requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
anyio==3.6.2
    # via starlette
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
commonmark==0.9.1
    # via rich
fastapi==0.86.0
    # via backupmanager (setup.py)
google-api-core[grpc]==2.10.2
    # via google-cloud-compute
google-auth==2.14.0
    # via google-api-core
google-cloud-compute==1.6.1
    # via backupmanager (setup.py)
googleapis-common-protos==1.56.4
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.50.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.50.0
    # via google-api-core
idna==3.4
    # via
    #   anyio
    #   requests
proto-plus==1.22.1
    # via google-cloud-compute
protobuf==4.21.9
    # via
    #   google-api-core
    #   google-cloud-compute
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydantic==1.10.2
    # via fastapi
pygments==2.13.0
    # via rich
requests==2.28.1
    # via google-api-core
rich==12.6.0
    # via backupmanager (setup.py)
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   grpcio
sniffio==1.3.0
    # via anyio
starlette==0.20.4
    # via fastapi
structlog==22.1.0
    # via backupmanager (setup.py)
typing-extensions==4.4.0
    # via pydantic
urllib3==1.26.12
    # via requests
